Id generating strategy migrated from Sequence to the new TableGenerator

+++ src/main/java/hello/repository/OrganizationRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Created by Xiaohu on 14-4-18.
  */
-public interface OrganizationRepository extends JpaRepository<Organization, Integer> {
+public interface OrganizationRepository extends JpaRepository<Organization, Long> {
 }
